Add MakersNewView.test.ts
